backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
plaid-python==15.0.0
python-dotenv==1.0.0
